Looks solid
Hi Jeff,

This is a good detailed proposal. You'll need to give special attention to defining and ordering the implementation of features (e.g.: how to decompose the problem and what to do first). You might consider at first producing a command line version of your application so that you can spend time in the most important part of the ISP (the algorithm). It would be straightforward to transition from a command-line-based program to providing an interface in the latter stages of your ISP.

Good work.

Mr. Gordon